+++ app.territoiresentransitions.react/src/app/pages/collectivite/CollectiviteRoutes.tsx
@@ -25,6 +25,12 @@ import {PersoReferentielPage} from './PersoReferentiel/PersoReferentielPage';
 import {PersoReferentielThematiquePage} from './PersoReferentielThematique/PersoReferentielThematiquePage';
 import {ParcoursLabellisationPage} from './ParcoursLabellisation/ParcoursLabellisationPage';
 import {ToutesLesCollectivitesPage} from '../ToutesLesCollectivites/ToutesLesCollectivitesPage';
+import {
+  CurrentCollectivite,
+  useCurrentCollectivite,
+} from 'core-logic/hooks/useCurrentCollectivite';
+import {RejoindreCetteCollectiviteDialog} from './RejoindreCetteCollectiviteDialog/RejoindreCetteCollectiviteDialog';
+import {getReferentContacts} from 'core-logic/api/procedures/collectiviteProcedures';
 
 /**
  * Routes starting with collectivite/:collectiviteId/ see App.ts Router.
@@ -33,8 +39,10 @@ import {ToutesLesCollectivitesPage} from '../ToutesLesCollectivites/ToutesLesCol
  */
 export const CollectiviteRoutes = () => {
   const {path} = useRouteMatch();
+  const currentCollectivite = useCurrentCollectivite();
   return (
     <>
+      <CollectiviteReadOnlyBanner collectivite={currentCollectivite} />
       <Route path={collectiviteReferentielPath}>
         <ReferentielsPage />
       </Route>
@@ -73,4 +81,22 @@ export const CollectiviteRoutes = () => {
       </Route>
     </>
   );
+};
+
+const CollectiviteReadOnlyBanner = ({
+  collectivite,
+}: {
+  collectivite: CurrentCollectivite | null;
+}) => {
+  if (!!collectivite && collectivite.readonly)
+    return (
+      <div className="flex justify-center items-center bg-yellow-400 py-4 bg-opacity-70">
+        <div className="text-sm mr-4">Lecture seule</div>
+        <RejoindreCetteCollectiviteDialog
+          getReferentContacts={getReferentContacts}
+          collectivite={collectivite}
+        />
+      </div>
+    );
+  return null;
 };
