+++ app.territoiresentransitions.react/src/app/pages/collectivite/RejoindreCetteCollectiviteDialog/RejoindreCetteCollectiviteDialog.stories.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import {Story, Meta} from '@storybook/react';
 import {
   RejoindreCetteCollectiviteDialog,
