+++ app.territoiresentransitions.react/src/app/pages/collectivite/RejoindreCetteCollectiviteDialog/RejoindreCetteCollectiviteDialog.tsx
@@ -1,4 +1,4 @@
-import {_RejoindreOuActiverDialogContent} from 'app/pages/MesCollectivites/_RejoindreOuActiverDialogContent';
+import {_RejoindreOuActiverDialogContent} from 'app/pages/collectivite/RejoindreCetteCollectiviteDialog/_RejoindreOuActiverDialogContent';
 import {ReferentContact} from 'core-logic/api/procedures/collectiviteProcedures';
 import {AllCollectiviteRead} from 'generated/dataLayer/all_collectivite_read';
 import React from 'react';
