+++ app.territoiresentransitions.react/src/app/pages/collectivite/RejoindreCetteCollectiviteDialog/_RejoindreOuActiverDialogContent.tsx
